Prescription & Invoice

+++ appointment/views.py
@@ -1,7 +1,7 @@
 from django.shortcuts import render, get_object_or_404, redirect
 from django.contrib.auth.decorators import login_required
 from accounts.models import Specialization, DoctorProfile, PatientProfile, Clinic, ClinicGallery, ReviewRating
-from .models import TimeSlot, PendingAppointment, Appointment
+from .models import TimeSlot, PendingAppointment, Appointment, Order
 from .forms import ReviewForm
 from datetime import datetime, timedelta
 from django.db.models import Q
@@ -44,6 +44,8 @@ def view_doctor(request,specialization_slug=None, doctor_slug=None):
 @login_required(login_url='login')
 def book_appointment(request,specialization_slug=None, doctor_slug=None):
     single_doctor = DoctorProfile.objects.get(specialization__slug=specialization_slug, slug=doctor_slug)
+    clinic = Clinic.objects.get(doctor=single_doctor)
+    reviews = ReviewRating.objects.filter(doctor=single_doctor)
     timeslots = TimeSlot.objects.filter(doctor=single_doctor, is_available=True)
     today = datetime.now().date()
     date_list = [today + timedelta(days=x) for x in range(7)]
@@ -51,6 +53,8 @@ def book_appointment(request,specialization_slug=None, doctor_slug=None):
         'single_doctor': single_doctor,
         'date_list': date_list,
         'timeslots': timeslots,
+        'clinic': clinic,
+        'reviews': reviews,
     }
     return  render(request, 'appointments/book_appointment.html', context)
 
@@ -59,8 +63,12 @@ def book_appointment(request,specialization_slug=None, doctor_slug=None):
 def checkout(request, doctor_slug=None):
     single_doctor = get_object_or_404(DoctorProfile, slug=doctor_slug)
     patientprofile = get_object_or_404(PatientProfile, user_id=request.user.id)
+    clinic = Clinic.objects.get(doctor=single_doctor)
+    reviews = ReviewRating.objects.filter(doctor=single_doctor)
     if request.method == 'POST':
         #user_form = UserForm(request.POST, instance=request.user)
+        tax = 0.02*single_doctor.price
+        total = tax+single_doctor.price
         selected_date = request.POST['selected_date']
         selected_timeslot = request.POST['selected_timeslot']
         # Get or create the PendingAppointment for the current user
@@ -82,29 +90,58 @@ def checkout(request, doctor_slug=None):
         'single_doctor': single_doctor,
         'selected_date': selected_date,
         'selected_timeslot': selected_timeslot,
+        'clinic': clinic,
+        'reviews': reviews,
+        'tax': tax,
+        'total': total,
     }
     return  render(request, 'appointments/checkout.html', context)
 
 
 @login_required(login_url='login')
 def booking_success(request):
-    patientprofile = get_object_or_404(PatientProfile, user_id=request.user.id)
-    pending_appointment = get_object_or_404(PendingAppointment, patient=patientprofile)
-    appointment = Appointment(
-        patient=patientprofile,
-        doctor=pending_appointment.doctor,
-        date=pending_appointment.date,
-        time=pending_appointment.time
-    )
-    appointment.save()
-    pending_appointment.delete()
-    appointment_date = appointment.date.strftime("%a")
-    timeslot = TimeSlot.objects.get(doctor=appointment.doctor, day=appointment_date, start_time=appointment.time)
-    timeslot.is_available = False
-    timeslot.save()
-    context={
-        'appointment': appointment,
-    }
+    if request.method == 'POST':
+        patientprofile = get_object_or_404(PatientProfile, user_id=request.user.id)
+        pending_appointment = get_object_or_404(PendingAppointment, patient=patientprofile)
+        appointment = Appointment(
+            patient=patientprofile,
+            doctor=pending_appointment.doctor,
+            date=pending_appointment.date,
+            time=pending_appointment.time,
+            type='Offline'
+        )
+        appointment.save()
+        pending_appointment.delete()
+        appointment_date = appointment.date.strftime("%a")
+        timeslot = TimeSlot.objects.get(doctor=appointment.doctor, day=appointment_date, start_time=appointment.time)
+        timeslot.is_available = False
+        timeslot.save()
+        first_name = request.POST.get('first_name')
+        last_name = request.POST.get('last_name')
+        email = request.POST.get('email')
+        phone = request.POST.get('phone')
+        billing_address = request.POST.get('billing_address')
+        payment_method = request.POST.get('payment_method')
+        tax = float(request.POST.get('tax'))
+        total_amount = float(request.POST.get('total_amount'))
+        order = Order(
+            first_name=first_name,
+            last_name=last_name,
+            email=email,
+            phone=phone,
+            billing_address=billing_address,
+            payment_method=payment_method,
+            doctor_profile=appointment.doctor,
+            patient_profile=appointment.patient,
+            tax=tax,
+            total_amount=total_amount,
+            description='Offline Appointment'
+        )
+        order.save()
+        context={
+            'appointment': appointment,
+            'order':order,
+        }
     return render(request, 'appointments/booking_success.html', context)
 
 
@@ -153,3 +190,14 @@ def submit_review(request, doctor_id):
                 data.save()
                 messages.success(request, 'Thank you! Your review has been submitted.')
                 return redirect(url)
+            
+
+@login_required(login_url='login')
+def view_invoice(request, order_id):
+    order = Order.objects.get(id=order_id)
+    amount = order.total_amount-order.tax
+    context = {
+        'order': order,
+        'amount': amount,
+    }
+    return  render(request, 'appointments/invoice.html', context)
